requirements.txt
chrisapp~=2.5.3
nose~=1.3.7
absl-py==1.0.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.0
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
nest-asyncio==1.5.4
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5